+++ Documentation/changelog.md
@@ -1,6 +1,7 @@
 ### Version 4.3.09_dev
 * Add Save in EEPROM ABL (Linear or Bilinear) or MBL
 * Add M323 Set Level bilinear manual - X<gridx> Y<gridy> Z<level val> S<level add>
+* Add Door open, a triggered door will prevent new commands from serial or sd card.
 * Fix HAL
 
 ### Version 4.3.08_dev